requirements.txt
openai
tavily-python
python-dotenv